requirements.txt
certifi==2018.1.18
chardet==3.0.4
diff-match-patch==20121119
Django==2.0.3
django-ckeditor==5.4.0
django-crispy-forms==1.7.2
django-formtools==2.1
django-import-export==1.0.0
django-js-asset==1.1.0
django-pure-pagination==0.3.0
django-ranged-response==0.2.0
django-reversion==2.0.13
django-simple-captcha==0.5.6
DjangoUeditor==1.8.143
et-xmlfile==1.0.1
future==0.16.0
httplib2==0.9.2
idna==2.6
jdcal==1.3
odfpy==1.3.6
openpyxl==2.5.1
Pillow==5.0.0
PyMySQL==0.8.0
pytz==2018.3
PyYAML==3.12
requests==2.18.4
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.22
xlrd==1.1.0
xlwt==1.3.0
